将所有已存在的im.txt进行规范化

+++ DLT/data/normalize.py
@@ -4,6 +4,19 @@ import numpy
 import optparse
 import os
 import shutil
+import pprint
+
+def walker(dir_path):
+	if len(dir_path) == 0:
+		dir_path = '.'
+
+	for root, dirs, files in os.walk(dir_path):
+		for name in files:
+			fname, ext = os.path.splitext(name)
+			if ext.lower() == '.ppm':
+				ppm_file = os.path.join(root, name)
+				update_coordinate(ppm_file)
+	
 
 def update_coordinate(image_path):
 	ppm = cv2.imread(image_path)
@@ -15,23 +28,25 @@ def update_coordinate(image_path):
 	new_txtfile = os.path.join(os.path.dirname(image_path), '%s_im_norm.txt' % fname)
 	print  '%s --> %s' % (txtfile, new_txtfile)
 
-	fp = open(txtfile, 'r')
-	tmp = open(new_txtfile, 'w')
+	if os.path.exists(txtfile):
+		fp = open(txtfile, 'r')
+		tmp = open(new_txtfile, 'w')
 
-	for line in fp:
-		x, y = line.strip().split()
-		x = float(x) *  iwidth / 500
-		y = float(y) * iheight / 500
-		x = int(round(x))
-		y = int(round(y))
-		tmp.write('%d %d\n' % (x, y))
+		for line in fp:
+			x, y = line.strip().split()
+			x = float(x) *  iwidth / 500
+			y = float(y) * iheight / 500
+			x = int(round(x))
+			y = int(round(y))
+			tmp.write('%d %d\n' % (x, y))
 
-	fp.close()
-	tmp.close()
+		fp.close()
+		tmp.close()
 if __name__ == '__main__':
 	parser = optparse.OptionParser()
 	parser.add_option('-i', '--image', dest='image', help="related image file")
 	(options, args) = parser.parse_args()
 
-	update_coordinate(options.image)
+	walker(os.path.dirname(options.image))
+	#update_coordinate(options.image)
